test(helm): separate `helm install` smoke test
