+++ test/e2e-helm/helm_suite_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2025.
+Copyright 2026.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -24,11 +24,17 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/emqx/emqx-operator/test/util"
+	. "github.com/emqx/emqx-operator/test/util"
 )
 
-// projectImage is the operator image which should already be built and available.
-var (
+const (
+	// helmReleaseName is the Helm release name used across all Helm upgrade tests.
+	helmReleaseName = "emqx-operator"
+
+	// localChartPath is the path to the local 2.3.0 chart being tested.
+	localChartPath = "deploy/charts/emqx-operator"
+
+	// operatorImage is the operator image which should already be built and available.
 	operatorImageRepo = "emqx/emqx-operator"
 	operatorImageTag  = "0.0.1-helm"
 	operatorImage     = operatorImageRepo + ":" + operatorImageTag
@@ -45,17 +51,41 @@ var _ = BeforeSuite(func() {
 	SetDefaultEventuallyPollingInterval(3 * time.Second)
 
 	By("add emqx Helm repository")
-	Expect(util.Run("helm", "repo", "add", "emqx", "https://repos.emqx.io/charts")).To(Succeed())
-	Expect(util.Run("helm", "repo", "update")).To(Succeed())
+	Expect(Run("helm", "repo", "add", "emqx", "https://repos.emqx.io/charts")).To(Succeed())
+	Expect(Run("helm", "repo", "update")).To(Succeed())
 
 	By("generate Helm chart files")
-	Expect(util.Run("make", "helm")).To(Succeed())
+	Expect(Run("make", "helm")).To(Succeed())
 
 	By("build emqx-operator docker image")
-	Expect(util.Run("make", "docker-build",
+	Expect(Run("make", "docker-build",
 		fmt.Sprintf("OPERATOR_IMAGE=%s", operatorImage),
 	)).To(Succeed())
 
 	By("load operator image into Kind cluster")
-	Expect(util.LoadImageToKindClusterWithName(operatorImage)).To(Succeed())
+	Expect(LoadImageToKindClusterWithName(operatorImage)).To(Succeed())
 })
+
+// crdExists checks whether a CRD exists in the cluster.
+func crdExists(name string) bool {
+	return Kubectl("get", "crd", name) == nil
+}
+
+// resourceExists checks whether a cluster-scoped resource exists.
+func resourceExists(kind, name string) bool {
+	return Kubectl("get", kind, name) == nil
+}
+
+// dumpHelmDiagnostics writes debug info for the given namespace to GinkgoWriter.
+func dumpHelmDiagnostics(namespace string) {
+	out, _ := Output("helm", "list", "--namespace", namespace)
+	GinkgoWriter.Print("Helm releases:\n", out)
+	out, _ = KubectlOut("get", "jobs", "--namespace", namespace, "-o", "yaml")
+	GinkgoWriter.Print("Jobs in namespace:\n", out)
+	out, _ = KubectlOut("get", "pods", "--namespace", namespace, "-o", "wide")
+	GinkgoWriter.Print("Pods in namespace:\n", out)
+	out, _ = KubectlOut("get", "crds", "-o", "custom-columns=NAME:.metadata.name,CONVERSION:.spec.conversion.strategy")
+	GinkgoWriter.Print("CRDs:\n", out)
+	out, _ = KubectlOut("get", "events", "--namespace", namespace, "--sort-by=.lastTimestamp")
+	GinkgoWriter.Print("Events:\n", out)
+}
